+++ src/NadekoBot/Migrations/NadekoContextModelSnapshot.cs
@@ -29,7 +29,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("ClubApplicants");
+                    b.ToTable("ClubApplicants", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Db.Models.ClubBans", b =>
@@ -44,7 +44,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("ClubBans");
+                    b.ToTable("ClubBans", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Db.Models.ClubInfo", b =>
@@ -86,7 +86,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("OwnerId")
                         .IsUnique();
 
-                    b.ToTable("Clubs");
+                    b.ToTable("Clubs", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Db.Models.DiscordUser", b =>
@@ -155,7 +155,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("DiscordUser");
+                    b.ToTable("DiscordUser", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Db.Models.FollowedStream", b =>
@@ -189,7 +189,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("FollowedStream");
+                    b.ToTable("FollowedStream", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AntiAltSetting", b =>
@@ -218,7 +218,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId")
                         .IsUnique();
 
-                    b.ToTable("AntiAltSetting");
+                    b.ToTable("AntiAltSetting", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AntiRaidSetting", b =>
@@ -250,7 +250,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId")
                         .IsUnique();
 
-                    b.ToTable("AntiRaidSetting");
+                    b.ToTable("AntiRaidSetting", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AntiSpamIgnore", b =>
@@ -272,7 +272,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("AntiSpamSettingId");
 
-                    b.ToTable("AntiSpamIgnore");
+                    b.ToTable("AntiSpamIgnore", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AntiSpamSetting", b =>
@@ -304,7 +304,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId")
                         .IsUnique();
 
-                    b.ToTable("AntiSpamSetting");
+                    b.ToTable("AntiSpamSetting", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AutoCommand", b =>
@@ -342,7 +342,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("AutoCommands");
+                    b.ToTable("AutoCommands", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AutoTranslateChannel", b =>
@@ -370,7 +370,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildId");
 
-                    b.ToTable("AutoTranslateChannels");
+                    b.ToTable("AutoTranslateChannels", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.AutoTranslateUser", b =>
@@ -398,7 +398,7 @@ namespace NadekoBot.Migrations
 
                     b.HasAlternateKey("ChannelId", "UserId");
 
-                    b.ToTable("AutoTranslateUsers");
+                    b.ToTable("AutoTranslateUsers", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.BanTemplate", b =>
@@ -421,7 +421,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId")
                         .IsUnique();
 
-                    b.ToTable("BanTemplates");
+                    b.ToTable("BanTemplates", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.BlacklistEntry", b =>
@@ -441,7 +441,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("Blacklist");
+                    b.ToTable("Blacklist", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.CommandAlias", b =>
@@ -466,7 +466,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("CommandAlias");
+                    b.ToTable("CommandAlias", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.CommandCooldown", b =>
@@ -491,7 +491,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("CommandCooldown");
+                    b.ToTable("CommandCooldown", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.CurrencyTransaction", b =>
@@ -529,7 +529,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("CurrencyTransactions");
+                    b.ToTable("CurrencyTransactions", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.DelMsgOnCmdChannel", b =>
@@ -554,7 +554,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("DelMsgOnCmdChannel");
+                    b.ToTable("DelMsgOnCmdChannel", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.DiscordPermOverride", b =>
@@ -580,7 +580,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId", "Command")
                         .IsUnique();
 
-                    b.ToTable("DiscordPermOverrides");
+                    b.ToTable("DiscordPermOverrides", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ExcludedItem", b =>
@@ -605,7 +605,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("XpSettingsId");
 
-                    b.ToTable("ExcludedItem");
+                    b.ToTable("ExcludedItem", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.FeedSub", b =>
@@ -631,7 +631,7 @@ namespace NadekoBot.Migrations
 
                     b.HasAlternateKey("GuildConfigId", "Url");
 
-                    b.ToTable("FeedSub");
+                    b.ToTable("FeedSub", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.FilterChannelId", b =>
@@ -653,7 +653,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("FilterChannelId");
+                    b.ToTable("FilterChannelId", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.FilteredWord", b =>
@@ -675,7 +675,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("FilteredWord");
+                    b.ToTable("FilteredWord", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.FilterLinksChannelId", b =>
@@ -697,7 +697,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("FilterLinksChannelId");
+                    b.ToTable("FilterLinksChannelId", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.FilterWordsChannelId", b =>
@@ -719,7 +719,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("FilterWordsChannelId");
+                    b.ToTable("FilterWordsChannelId", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.GCChannelId", b =>
@@ -741,7 +741,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("GCChannelId");
+                    b.ToTable("GCChannelId", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.GroupName", b =>
@@ -767,7 +767,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId", "Number")
                         .IsUnique();
 
-                    b.ToTable("GroupName");
+                    b.ToTable("GroupName", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.GuildConfig", b =>
@@ -897,7 +897,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("WarnExpireHours");
 
-                    b.ToTable("GuildConfigs");
+                    b.ToTable("GuildConfigs", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.IgnoredLogItem", b =>
@@ -923,7 +923,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("LogSettingId", "LogItemId", "ItemType")
                         .IsUnique();
 
-                    b.ToTable("IgnoredLogChannels");
+                    b.ToTable("IgnoredLogChannels", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.IgnoredVoicePresenceChannel", b =>
@@ -945,7 +945,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("LogSettingId");
 
-                    b.ToTable("IgnoredVoicePresenceCHannels");
+                    b.ToTable("IgnoredVoicePresenceCHannels", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ImageOnlyChannel", b =>
@@ -968,7 +968,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("ChannelId")
                         .IsUnique();
 
-                    b.ToTable("ImageOnlyChannels");
+                    b.ToTable("ImageOnlyChannels", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.LogSetting", b =>
@@ -1033,7 +1033,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId")
                         .IsUnique();
 
-                    b.ToTable("LogSettings");
+                    b.ToTable("LogSettings", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.MusicPlayerSettings", b =>
@@ -1067,7 +1067,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId")
                         .IsUnique();
 
-                    b.ToTable("MusicPlayerSettings");
+                    b.ToTable("MusicPlayerSettings", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.MusicPlaylist", b =>
@@ -1090,7 +1090,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("MusicPlaylists");
+                    b.ToTable("MusicPlaylists", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.MutedUserId", b =>
@@ -1112,7 +1112,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("MutedUserId");
+                    b.ToTable("MutedUserId", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.NadekoExpression", b =>
@@ -1150,7 +1150,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("Expressions");
+                    b.ToTable("Expressions", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.NsfwBlacklistedTag", b =>
@@ -1172,7 +1172,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildId");
 
-                    b.ToTable("NsfwBlacklistedTags");
+                    b.ToTable("NsfwBlacklistedTags", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Permissionv2", b =>
@@ -1212,7 +1212,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("Permissions");
+                    b.ToTable("Permissions", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.PlantedCurrency", b =>
@@ -1249,7 +1249,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("MessageId")
                         .IsUnique();
 
-                    b.ToTable("PlantedCurrency");
+                    b.ToTable("PlantedCurrency", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.PlaylistSong", b =>
@@ -1283,7 +1283,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("MusicPlaylistId");
 
-                    b.ToTable("PlaylistSong");
+                    b.ToTable("PlaylistSong", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Poll", b =>
@@ -1309,7 +1309,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId")
                         .IsUnique();
 
-                    b.ToTable("Poll");
+                    b.ToTable("Poll", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.PollAnswer", b =>
@@ -1334,7 +1334,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("PollId");
 
-                    b.ToTable("PollAnswer");
+                    b.ToTable("PollAnswer", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.PollVote", b =>
@@ -1359,7 +1359,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("PollId");
 
-                    b.ToTable("PollVote");
+                    b.ToTable("PollVote", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Quote", b =>
@@ -1395,7 +1395,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("Keyword");
 
-                    b.ToTable("Quotes");
+                    b.ToTable("Quotes", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ReactionRole", b =>
@@ -1420,7 +1420,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("ReactionRoleMessageId");
 
-                    b.ToTable("ReactionRole");
+                    b.ToTable("ReactionRole", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ReactionRoleMessage", b =>
@@ -1451,7 +1451,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("ReactionRoleMessage");
+                    b.ToTable("ReactionRoleMessage", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Reminder", b =>
@@ -1485,7 +1485,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("When");
 
-                    b.ToTable("Reminders");
+                    b.ToTable("Reminders", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Repeater", b =>
@@ -1520,7 +1520,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("Repeaters");
+                    b.ToTable("Repeaters", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.RewardedUser", b =>
@@ -1549,7 +1549,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("PatreonUserId")
                         .IsUnique();
 
-                    b.ToTable("RewardedUsers");
+                    b.ToTable("RewardedUsers", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.RotatingPlayingStatus", b =>
@@ -1569,7 +1569,7 @@ namespace NadekoBot.Migrations
 
                     b.HasKey("Id");
 
-                    b.ToTable("RotatingStatus");
+                    b.ToTable("RotatingStatus", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.SelfAssignedRole", b =>
@@ -1600,7 +1600,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildId", "RoleId")
                         .IsUnique();
 
-                    b.ToTable("SelfAssignableRoles");
+                    b.ToTable("SelfAssignableRoles", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ShopEntry", b =>
@@ -1640,7 +1640,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("ShopEntry");
+                    b.ToTable("ShopEntry", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.ShopEntryItem", b =>
@@ -1662,7 +1662,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("ShopEntryId");
 
-                    b.ToTable("ShopEntryItem");
+                    b.ToTable("ShopEntryItem", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.SlowmodeIgnoredRole", b =>
@@ -1684,7 +1684,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("SlowmodeIgnoredRole");
+                    b.ToTable("SlowmodeIgnoredRole", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.SlowmodeIgnoredUser", b =>
@@ -1706,7 +1706,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("SlowmodeIgnoredUser");
+                    b.ToTable("SlowmodeIgnoredUser", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.StreamRoleBlacklistedUser", b =>
@@ -1731,7 +1731,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("StreamRoleSettingsId");
 
-                    b.ToTable("StreamRoleBlacklistedUser");
+                    b.ToTable("StreamRoleBlacklistedUser", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.StreamRoleSettings", b =>
@@ -1763,7 +1763,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId")
                         .IsUnique();
 
-                    b.ToTable("StreamRoleSettings");
+                    b.ToTable("StreamRoleSettings", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.StreamRoleWhitelistedUser", b =>
@@ -1788,7 +1788,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("StreamRoleSettingsId");
 
-                    b.ToTable("StreamRoleWhitelistedUser");
+                    b.ToTable("StreamRoleWhitelistedUser", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.UnbanTimer", b =>
@@ -1813,7 +1813,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("UnbanTimer");
+                    b.ToTable("UnbanTimer", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.UnmuteTimer", b =>
@@ -1838,7 +1838,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("UnmuteTimer");
+                    b.ToTable("UnmuteTimer", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.UnroleTimer", b =>
@@ -1866,7 +1866,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("UnroleTimer");
+                    b.ToTable("UnroleTimer", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.UserXpStats", b =>
@@ -1911,7 +1911,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("UserId", "GuildId")
                         .IsUnique();
 
-                    b.ToTable("UserXpStats");
+                    b.ToTable("UserXpStats", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.VcRoleInfo", b =>
@@ -1936,7 +1936,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("VcRoleInfo");
+                    b.ToTable("VcRoleInfo", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.WaifuInfo", b =>
@@ -1971,7 +1971,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("WaifuId")
                         .IsUnique();
 
-                    b.ToTable("WaifuInfo");
+                    b.ToTable("WaifuInfo", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.WaifuItem", b =>
@@ -1996,7 +1996,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("WaifuInfoId");
 
-                    b.ToTable("WaifuItem");
+                    b.ToTable("WaifuItem", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.WaifuUpdate", b =>
@@ -2028,7 +2028,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("WaifuUpdates");
+                    b.ToTable("WaifuUpdates", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.Warning", b =>
@@ -2071,7 +2071,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("UserId");
 
-                    b.ToTable("Warnings");
+                    b.ToTable("Warnings", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.WarningPunishment", b =>
@@ -2102,7 +2102,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("GuildConfigId");
 
-                    b.ToTable("WarningPunishment");
+                    b.ToTable("WarningPunishment", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.XpCurrencyReward", b =>
@@ -2127,7 +2127,7 @@ namespace NadekoBot.Migrations
 
                     b.HasIndex("XpSettingsId");
 
-                    b.ToTable("XpCurrencyReward");
+                    b.ToTable("XpCurrencyReward", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.XpRoleReward", b =>
@@ -2156,7 +2156,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("XpSettingsId", "Level")
                         .IsUnique();
 
-                    b.ToTable("XpRoleReward");
+                    b.ToTable("XpRoleReward", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Services.Database.Models.XpSettings", b =>
@@ -2179,7 +2179,7 @@ namespace NadekoBot.Migrations
                     b.HasIndex("GuildConfigId")
                         .IsUnique();
 
-                    b.ToTable("XpSettings");
+                    b.ToTable("XpSettings", (string)null);
                 });
 
             modelBuilder.Entity("NadekoBot.Db.Models.ClubApplicants", b =>